refactor: convert omnichannel callbacks to ts (#28564)
