--- apps/meteor/ee/app/livechat-enterprise/server/hooks/onLoadConfigApi.js
@@ -1,23 +1,0 @@
-import { callbacks } from '../../../../../lib/callbacks';
-import { getLivechatQueueInfo, getLivechatCustomFields } from '../lib/Helper';
-
-const onLoadConfigApi = async (options = {}) => {
-	const { room } = options;
-
-	const queueInfo = await getLivechatQueueInfo(room);
-	const customFields = await getLivechatCustomFields();
-
-	return {
-		...(queueInfo && { queueInfo }),
-		...(customFields && { customFields }),
-		...options,
-	};
-};
-
-callbacks.add(
-	'livechat.onLoadConfigApi',
-	// TODO callbacks cannot be async
-	onLoadConfigApi,
-	callbacks.priority.MEDIUM,
-	'livechat-on-load-config-api',
-);
