+++ apps/meteor/lib/callbacks.ts
@@ -38,7 +38,8 @@ enum CallbackPriority {
  *
  * TODO: move those to event-based systems
  */
-type EventLikeCallbackSignatures = {
+// eslint-disable-next-line @typescript-eslint/naming-convention
+interface EventLikeCallbackSignatures {
 	'afterActivateUser': (user: IUser) => void;
 	'afterCreateChannel': (owner: IUser, room: IRoom) => void;
 	'afterCreatePrivateGroup': (owner: IUser, room: IRoom) => void;
@@ -59,7 +60,7 @@ type EventLikeCallbackSignatures = {
 	'livechat:afterReturnRoomAsInquiry': (params: { room: IRoom }) => void;
 	'livechat.setUserStatusLivechat': (params: { userId: IUser['_id']; status: OmnichannelAgentStatus }) => void;
 	'livechat.agentStatusChanged': (params: { userId: IUser['_id']; status: OmnichannelAgentStatus }) => void;
-	'livechat.afterTakeInquiry': (inq: ILivechatInquiryRecord, agent: ILivechatAgent) => void;
+	'livechat.afterTakeInquiry': (inq: ILivechatInquiryRecord, agent: { agentId: string; username: string }) => void;
 	'afterAddedToRoom': (params: { user: IUser; inviter?: IUser }, room: IRoom) => void;
 	'beforeAddedToRoom': (params: { user: IUser; inviter: IUser }) => void;
 	'afterCreateDirectRoom': (params: IRoom, second: { members: IUser[]; creatorId: IUser['_id'] }) => void;
@@ -86,7 +87,7 @@ type EventLikeCallbackSignatures = {
 	'afterMuteUser': (users: { mutedUser: IUser; fromUser: IUser }, room: IRoom) => void;
 	'beforeUnmuteUser': (users: { mutedUser: IUser; fromUser: IUser }, room: IRoom) => void;
 	'afterUnmuteUser': (users: { mutedUser: IUser; fromUser: IUser }, room: IRoom) => void;
-};
+}
 
 /**
  * Callbacks that are supposed to be composed like a chain.
@@ -94,6 +95,33 @@ type EventLikeCallbackSignatures = {
  * TODO: develop a middleware alternative and grant independence of execution order
  */
 type ChainedCallbackSignatures = {
+	'livechat.beforeRoom': (
+		roomInfo: Record<string, unknown>,
+		extraData?: Record<string, unknown> & { sla?: string },
+	) => Record<string, unknown>;
+	'livechat.newRoom': (room: IOmnichannelRoom) => IOmnichannelRoom;
+
+	'livechat.beforeForwardRoomToDepartment': <T extends { room: IOmnichannelRoom; transferData?: { department: { _id: string } } }>(
+		options: T,
+	) => T;
+
+	'livechat.beforeRouteChat': (inquiry: ILivechatInquiryRecord, agent?: { agentId: string; username: string }) => ILivechatInquiryRecord;
+	'livechat.checkDefaultAgentOnNewRoom': (
+		agent: { agentId: string; username: string },
+		visitor?: ILivechatVisitor,
+	) => { agentId: string; username: string };
+
+	'livechat.onLoadForwardDepartmentRestrictions': (params: { departmentId: string }) => Record<string, unknown>;
+
+	'livechat.saveInfo': (
+		newRoom: IOmnichannelRoom,
+		props: { user: Required<Pick<IUser, '_id' | 'username' | 'name'>>; oldRoom: IOmnichannelRoom },
+	) => IOmnichannelRoom;
+
+	'livechat.onCheckRoomApiParams': (params: Record<string, unknown>) => Record<string, unknown>;
+
+	'livechat.onLoadConfigApi': (config: { room: IOmnichannelRoom }) => Record<string, unknown>;
+
 	'beforeSaveMessage': (message: IMessage, room?: IRoom) => IMessage;
 	'afterCreateUser': (user: IUser) => IUser;
 	'afterDeleteRoom': (rid: IRoom['_id']) => IRoom['_id'];
@@ -133,7 +161,6 @@ type ChainedCallbackSignatures = {
 	'on-business-hour-start': (params: { BusinessHourBehaviorClass: { new (): IBusinessHourBehavior } }) => {
 		BusinessHourBehaviorClass: { new (): IBusinessHourBehavior };
 	};
-	'livechat.saveInfo': (newRoom: IOmnichannelRoom, { user, oldRoom }: { user: IUser; oldRoom: IOmnichannelRoom }) => IOmnichannelRoom;
 	'renderMessage': <T extends IMessage & { html: string }>(message: T) => T;
 	'oembed:beforeGetUrlContent': (data: {
 		urlObj: Omit<UrlWithParsedQuery, 'host' | 'search'> & { host?: unknown; search?: unknown };
@@ -188,17 +215,13 @@ type Hook =
 	| 'livechat.beforeRouteChat'
 	| 'livechat.chatQueued'
 	| 'livechat.checkAgentBeforeTakeInquiry'
-	| 'livechat.checkDefaultAgentOnNewRoom'
 	| 'livechat.sendTranscript'
 	| 'livechat.closeRoom'
 	| 'livechat.leadCapture'
-	| 'livechat.newRoom'
 	| 'livechat.offlineMessage'
 	| 'livechat.onAgentAssignmentFailed'
 	| 'livechat.onCheckRoomApiParams'
 	| 'livechat.onLoadConfigApi'
-	| 'livechat.onLoadForwardDepartmentRestrictions'
-	| 'livechat.saveInfo'
 	| 'loginPageStateChange'
 	| 'mapLDAPUserData'
 	| 'onCreateUser'
